refactor(acp): extract shared content-validation module (#3896)
Extract validatePromptOutput, extractResultText, HALLUCINATION_SIGNATURES,
and PromptValidationWarning into src/services/acp/content-validation.ts.

Previously duplicated between backend.ts and the test file — tests validated
their own copy, not production code. Now both import from the shared module.

Added 7 new unit tests for extractResultText and HALLUCINATION_SIGNATURES.

Fixes: https://github.com/OneStepAt4time/aegis/issues/3896
